Refactor core swarm functions
